tmp: Pipeline adoption started
