+++ pkg/dbsp/document.go
@@ -2,45 +2,13 @@ package dbsp
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 )
 
 // Document represents an unstructured document as map[string]any
 // Can contain embedded maps, slices, and primitives (int64, float64, string, bool)
 type Document = map[string]any
-
-// DocumentZSet implements Z-sets for atomic documents
-// Documents are treated as opaque units - no internal structure operations
-type DocumentZSet struct {
-	// Use JSON representation as key since documents aren't directly comparable
-	docs   map[string]Document // JSON key -> original document
-	counts map[string]int      // JSON key -> multiplicity
-}
-
-// Error type for better error handling
-type ZSetError struct {
-	Message string
-	Cause   error
-}
-
-func (e *ZSetError) Error() string {
-	if e.Cause != nil {
-		return fmt.Sprintf("%s: %v", e.Message, e.Cause)
-	}
-	return e.Message
-}
-
-func newZSetError(message string, cause error) error {
-	return &ZSetError{Message: message, Cause: cause}
-}
-
-// NewDocumentZSet creates an empty DocumentZSet
-func NewDocumentZSet() *DocumentZSet {
-	return &DocumentZSet{
-		docs:   make(map[string]Document),
-		counts: make(map[string]int),
-	}
-}
 
 // computeJSONKey creates a deterministic JSON representation for document identity
 // This is the key function that defines document equality
@@ -147,6 +115,19 @@ func deepCopy(val any) (any, error) {
 	}
 }
 
+// deepCopyDocument creates a deep copy of a document.
+func deepCopyDocument(val any) (Document, error) {
+	c, err := deepCopy(val)
+	if err != nil {
+		return nil, err
+	}
+	doc, ok := c.(Document)
+	if !ok {
+		return nil, errors.New("failed to cast document after deepCopy")
+	}
+	return doc, nil
+}
+
 // computeJSONAny creates a deterministic JSON representation for an arbitrary any value
 func computeJSONAny(doc any) (string, error) {
 	bytes, err := json.Marshal(doc)
@@ -155,324 +136,4 @@ func computeJSONAny(doc any) (string, error) {
 	}
 
 	return string(bytes), nil
-}
-
-// AddDocument adds a document to the ZSet with given multiplicity by modifying the Zset in
-// place. This is the core operation for building Z-sets.
-func (dz *DocumentZSet) AddDocumentMutate(doc Document, count int) error {
-	if count == 0 {
-		return nil
-	}
-
-	key, err := computeJSONKey(doc)
-	if err != nil {
-		return err
-	}
-
-	if _, exists := dz.counts[key]; exists {
-		dz.counts[key] += count
-	} else {
-		copied, err := deepCopy(doc)
-		if err != nil {
-			return err
-		}
-		dz.docs[key] = copied.(Document)
-		dz.counts[key] = count
-	}
-
-	if dz.counts[key] == 0 {
-		delete(dz.counts, key)
-		delete(dz.docs, key)
-	}
-
-	return nil
-}
-
-// AddDocument adds a document to the ZSet with given multiplicity by creating a new ZSet. This is
-// the core operation for building Z-sets.
-func (dz *DocumentZSet) AddDocument(doc Document, count int) (*DocumentZSet, error) {
-	result, err := dz.DeepCopy()
-	if err != nil {
-		return nil, err
-	}
-	err = result.AddDocumentMutate(doc, count)
-	return result, err
-}
-
-// Add performs Z-set addition (union with multiplicity)
-func (dz *DocumentZSet) Add(other *DocumentZSet) (*DocumentZSet, error) {
-	if other == nil {
-		return dz.DeepCopy()
-	}
-
-	result, err := dz.DeepCopy()
-	if err != nil {
-		return nil, newZSetError("failed to copy Z-set for addition", err)
-	}
-
-	// Add each document from other Z-set
-	for key, count := range other.counts {
-		doc := other.docs[key]
-		if err := result.AddDocumentMutate(doc, count); err != nil {
-			return nil, newZSetError("failed to add document during Z-set addition", err)
-		}
-	}
-
-	return result, nil
-}
-
-// Subtract performs Z-set subtraction
-func (dz *DocumentZSet) Subtract(other *DocumentZSet) (*DocumentZSet, error) {
-	if other == nil {
-		return dz.DeepCopy()
-	}
-
-	result, err := dz.DeepCopy()
-	if err != nil {
-		return nil, newZSetError("failed to copy Z-set for subtraction", err)
-	}
-
-	// Subtract each document from other Z-set (add with negative count)
-	for key, count := range other.counts {
-		doc := other.docs[key]
-		if err := result.AddDocumentMutate(doc, -count); err != nil {
-			return nil, newZSetError("failed to subtract document during Z-set subtraction", err)
-		}
-	}
-
-	return result, nil
-}
-
-// Distinct converts Z-set to set semantics (all multiplicities become 1)
-// This is crucial for converting from multiset to set semantics
-func (dz *DocumentZSet) Distinct() (*DocumentZSet, error) {
-	result := NewDocumentZSet()
-
-	for key, count := range dz.counts {
-		if count > 0 {
-			doc := dz.docs[key]
-			var err error
-			result, err = result.AddDocument(doc, 1) // Force count to 1
-			if err != nil {
-				return nil, newZSetError("failed to add document during distinct operation", err)
-			}
-		}
-		// Note: negative counts are filtered out (don't appear in result)
-	}
-
-	return result, nil
-}
-
-// DeepCopy creates a deep copy of the DocumentZSet
-func (dz *DocumentZSet) DeepCopy() (*DocumentZSet, error) {
-	result := &DocumentZSet{
-		docs:   make(map[string]Document),
-		counts: make(map[string]int),
-	}
-
-	for key, doc := range dz.docs {
-		copied, err := deepCopy(doc)
-		if err != nil {
-			return nil, newZSetError("failed to deep copy document in Z-set", err)
-		}
-		result.docs[key] = copied.(Document)
-		result.counts[key] = dz.counts[key]
-	}
-
-	return result, nil
-}
-
-// DocumentEntry represents a document with its multiplicity in a Z-set.
-type DocumentEntry struct {
-	Document     Document
-	Multiplicity int
-}
-
-// List returns all documents with their multiplicities (including negative ones).
-func (dz *DocumentZSet) List() ([]DocumentEntry, error) {
-	result := make([]DocumentEntry, 0, len(dz.counts))
-
-	for key, mult := range dz.counts {
-		doc, err := deepCopy(dz.docs[key])
-		if err != nil {
-			return nil, newZSetError("failed to copy document in List", err)
-		}
-
-		result = append(result, DocumentEntry{
-			Document:     doc.(Document),
-			Multiplicity: mult,
-		})
-	}
-
-	return result, nil
-}
-
-// GetDocuments returns all documents as a slice (with multiplicities)
-// Documents with multiplicity n appear n times in the result
-func (dz *DocumentZSet) GetDocuments() ([]Document, error) {
-	var result []Document
-
-	for key, count := range dz.counts {
-		if count <= 0 {
-			continue // Skip non-positive counts
-		}
-
-		doc := dz.docs[key]
-		for i := 0; i < count; i++ {
-			copied, err := deepCopy(doc)
-			if err != nil {
-				return nil, newZSetError("failed to copy document in GetDocuments", err)
-			}
-			result = append(result, copied.(Document))
-		}
-	}
-
-	return result, nil
-}
-
-// GetUniqueDocuments returns all unique documents (ignoring multiplicities)
-func (dz *DocumentZSet) GetUniqueDocuments() ([]Document, error) {
-	var result []Document
-
-	for key, count := range dz.counts {
-		if count > 0 {
-			doc := dz.docs[key]
-			copied, err := deepCopy(doc)
-			if err != nil {
-				return nil, newZSetError("failed to copy document in GetUniqueDocuments", err)
-			}
-			result = append(result, copied.(Document))
-		}
-	}
-
-	return result, nil
-}
-
-// IsZero checks if the Z-set is empty (no documents with positive multiplicity)
-func (dz *DocumentZSet) IsZero() bool {
-	return len(dz.counts) == 0
-}
-
-// Size returns the number of documents counting only positive multiplicities.
-func (dz *DocumentZSet) Size() int {
-	total := 0
-	for _, count := range dz.counts {
-		if count > 0 {
-			total += count
-		}
-	}
-	return total
-}
-
-// TotalSize returns the total number of documents, counting both positive and negative multiplicities.
-func (dz *DocumentZSet) TotalSize() int {
-	total := 0
-	for _, count := range dz.counts {
-		if count > 0 {
-			total += count
-		} else {
-			total += -count
-		}
-	}
-	return total
-}
-
-// UniqueCount returns number of unique documents (ignoring multiplicities)
-func (dz *DocumentZSet) UniqueCount() int {
-	count := 0
-	for _, multiplicity := range dz.counts {
-		if multiplicity > 0 {
-			count++
-		}
-	}
-	return count
-}
-
-// GetMultiplicity returns the multiplicity of a specific document
-func (dz *DocumentZSet) GetMultiplicity(doc Document) (int, error) {
-	key, err := computeJSONKey(doc)
-	if err != nil {
-		return 0, newZSetError("failed to compute document key", err)
-	}
-
-	if count, exists := dz.counts[key]; exists {
-		return count, nil
-	}
-
-	return 0, nil // Document not in Z-set
-}
-
-// Contains checks if a document exists in the Z-set (with positive multiplicity)
-func (dz *DocumentZSet) Contains(doc Document) (bool, error) {
-	multiplicity, err := dz.GetMultiplicity(doc)
-	if err != nil {
-		return false, err
-	}
-	return multiplicity > 0, nil
-}
-
-// Helper functions for creating documents and Z-sets
-func NewDocument() Document { return make(Document) }
-
-// newDocumentFromPairs creates a new document from key-value pairs
-func newDocumentFromPairs(pairs ...any) (Document, error) {
-	if len(pairs)%2 != 0 {
-		return nil, newZSetError("NewDocument requires even number of arguments (key-value pairs)", nil)
-	}
-
-	doc := make(Document)
-	for i := 0; i < len(pairs); i += 2 {
-		key, ok := pairs[i].(string)
-		if !ok {
-			return nil, newZSetError(fmt.Sprintf("key at position %d must be a string", i), nil)
-		}
-		value := pairs[i+1]
-		doc[key] = value
-	}
-
-	return doc, nil
-}
-
-// SingletonZSet creates a Z-set containing a single document with multiplicity 1
-func SingletonZSet(doc Document) (*DocumentZSet, error) {
-	zset := NewDocumentZSet()
-	return zset.AddDocument(doc, 1)
-}
-
-// FromDocuments creates a Z-set from a slice of documents (each with multiplicity 1)
-func FromDocuments(docs []Document) (*DocumentZSet, error) {
-	result := NewDocumentZSet()
-
-	for i, doc := range docs {
-		var err error
-		result, err = result.AddDocument(doc, 1)
-		if err != nil {
-			return nil, newZSetError(fmt.Sprintf("failed to add document at index %d", i), err)
-		}
-	}
-
-	return result, nil
-}
-
-// String returns a string representation of the Z-set for debugging
-func (dz *DocumentZSet) String() string {
-	if dz.IsZero() {
-		return "∅"
-	}
-
-	result := "{"
-	first := true
-
-	for key, count := range dz.counts {
-		if !first {
-			result += ", "
-		}
-
-		doc := dz.docs[key]
-		result += fmt.Sprintf("%v×%d", doc, count)
-		first = false
-	}
-
-	result += "}"
-	return result
 }
